pulled out some classes and functions into separate files
